@@ -4,4 +4,4 @@
 
 module github.com/raohwork/jsonapi
 
-go 1.12
+go 1.21
